+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   "decorator": "3.20.11",
   "core": "3.20.11",
-  "timestamp": 1758766383361
+  "timestamp": 1759896568498
 };