+++ scripts/rollback/rollback-3.20.16.sh
@@ -1,0 +1,72 @@
+#!/bin/bash
+# timestamp: 1759896490377
+
+npm dist-tag add @midwayjs/bull-board@3.20.15 latest
+npm dist-tag add @midwayjs/busboy@3.20.15 latest
+npm dist-tag add @midwayjs/captcha@3.20.15 latest
+npm dist-tag add @midwayjs/casbin@3.20.15 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.20.15 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.20.15 latest
+npm dist-tag add @midwayjs/code-dye@3.20.15 latest
+npm dist-tag add @midwayjs/consul@3.20.15 latest
+npm dist-tag add @midwayjs/cross-domain@3.20.15 latest
+npm dist-tag add @midwayjs/http-proxy@3.20.15 latest
+npm dist-tag add @midwayjs/i18n@3.20.15 latest
+npm dist-tag add @midwayjs/info@3.20.15 latest
+npm dist-tag add @midwayjs/jwt@3.20.15 latest
+npm dist-tag add @midwayjs/leoric@3.20.15 latest
+npm dist-tag add @midwayjs/mikro@3.20.15 latest
+npm dist-tag add @midwayjs/otel@3.20.15 latest
+npm dist-tag add @midwayjs/passport@3.20.15 latest
+npm dist-tag add @midwayjs/process-agent@3.20.15 latest
+npm dist-tag add @midwayjs/prometheus@3.20.15 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.20.15 latest
+npm dist-tag add @midwayjs/security@3.20.15 latest
+npm dist-tag add @midwayjs/sequelize@3.20.15 latest
+npm dist-tag add @midwayjs/static-file@3.20.15 latest
+npm dist-tag add @midwayjs/swagger@3.20.15 latest
+npm dist-tag add @midwayjs/tags@3.20.15 latest
+npm dist-tag add @midwayjs/upload@3.20.15 latest
+npm dist-tag add @midwayjs/validate@3.20.15 latest
+npm dist-tag add @midwayjs/version@3.20.15 latest
+npm dist-tag add @midwayjs/view@3.20.15 latest
+npm dist-tag add @midwayjs/view-ejs@3.20.15 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.20.15 latest
+npm dist-tag add @midwayjs/koa@3.20.15 latest
+npm dist-tag add @midwayjs/ws@3.20.15 latest
+
+# Changes:
+
+#  - @midwayjs/bull-board: 3.20.15 => 3.20.16
+#  - @midwayjs/busboy: 3.20.15 => 3.20.16
+#  - @midwayjs/captcha: 3.20.15 => 3.20.16
+#  - @midwayjs/casbin: 3.20.15 => 3.20.16
+#  - @midwayjs/casbin-redis-adapter: 3.20.15 => 3.20.16
+#  - @midwayjs/casbin-typeorm-adapter: 3.20.15 => 3.20.16
+#  - @midwayjs/code-dye: 3.20.15 => 3.20.16
+#  - @midwayjs/consul: 3.20.15 => 3.20.16
+#  - @midwayjs/cross-domain: 3.20.15 => 3.20.16
+#  - @midwayjs/http-proxy: 3.20.15 => 3.20.16
+#  - @midwayjs/i18n: 3.20.15 => 3.20.16
+#  - @midwayjs/info: 3.20.15 => 3.20.16
+#  - @midwayjs/jwt: 3.20.15 => 3.20.16
+#  - @midwayjs/leoric: 3.20.15 => 3.20.16
+#  - @midwayjs/mikro: 3.20.15 => 3.20.16
+#  - @midwayjs/otel: 3.20.15 => 3.20.16
+#  - @midwayjs/passport: 3.20.15 => 3.20.16
+#  - @midwayjs/process-agent: 3.20.15 => 3.20.16
+#  - @midwayjs/prometheus: 3.20.15 => 3.20.16
+#  - @midwayjs/prometheus-socket-io: 3.20.15 => 3.20.16
+#  - @midwayjs/security: 3.20.15 => 3.20.16
+#  - @midwayjs/sequelize: 3.20.15 => 3.20.16
+#  - @midwayjs/static-file: 3.20.15 => 3.20.16
+#  - @midwayjs/swagger: 3.20.15 => 3.20.16
+#  - @midwayjs/tags: 3.20.15 => 3.20.16
+#  - @midwayjs/upload: 3.20.15 => 3.20.16
+#  - @midwayjs/validate: 3.20.15 => 3.20.16
+#  - @midwayjs/version: 3.20.15 => 3.20.16
+#  - @midwayjs/view: 3.20.15 => 3.20.16
+#  - @midwayjs/view-ejs: 3.20.15 => 3.20.16
+#  - @midwayjs/view-nunjucks: 3.20.15 => 3.20.16
+#  - @midwayjs/koa: 3.20.15 => 3.20.16
+#  - @midwayjs/ws: 3.20.15 => 3.20.16